backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
pydantic==2.5.3
email-validator==2.1.0
certifi==2024.8.30
